rename widget files
